Revert "merged updated build.gradle for docker compose v2 usage"
This reverts commit f6d91113d8de05f205c14b0f47c948ba94ec8279, reversing
changes made to 8cda4c37a51b126b54e064a2166a7ce59bb2734c.
