--- app/gzac/src/main/resources/config/documenten-api/bezwaar.zgw-document-list-column.json
@@ -1,29 +1,0 @@
-{
-    "changesetId": "bezwaar.zgw-document-list-columns-v1",
-    "case-definitions": [
-        {
-            "key": "bezwaar",
-            "columns": [
-                {
-                    "key": "TITEL"
-                },
-                {
-                    "key": "BESTANDSNAAM"
-                },
-                {
-                    "key": "CREATIEDATUM",
-                    "defaultSort": "DESC"
-                },
-                {
-                    "key": "STATUS"
-                },
-                {
-                    "key": "VERSIE"
-                },
-                {
-                    "key": "INFORMATIEOBJECTTYPE"
-                }
-            ]
-        }
-    ]
-}
