+++ build.gradle
@@ -71,8 +71,7 @@ sonar {
 }
 
 dockerCompose {
-    useDockerComposeV2 = true
-    removeOrphans = true
+    useDockerComposeV2 = false
 }
 
 configurations {
@@ -160,7 +159,7 @@ subprojects { subproject ->
         }
 
         dockerCompose {
-            useDockerComposeV2 = true
+            useDockerComposeV2 = false
         }
 
         tasks.withType(KotlinCompile).configureEach {
