+++ gradle.properties
@@ -20,8 +20,6 @@ namedVolumeNameTest=valtimo-test-fixtures
 org.gradle.daemon=true
 org.gradle.jvmargs=-Xmx6g -XX:MaxMetaspaceSize=2048m -XX:+HeapDumpOnOutOfMemoryError
 org.gradle.parallel=false
-https.socketTimeout=60000
-https.connectionTimeout=60000
 # override liquibase to fix bug see :
 # https://forum.liquibase.org/t/liquibaseexception-java-lang-classcastexception-class-java-time-localdatetime-cannot-be-cast-to-class-java-lang-string/5059/3
 
@@ -101,4 +99,4 @@ snakeYamlVersionOverride=2.2
 
 # project version
 
-projectVersion=12.1.0
+projectVersion=12.0.0
