--- plugin-authentication/openzaak-plugin-authentication/src/main/resources/config/documenten-api/maykin-1.13.0.zgw-documenten-api-version.json
@@ -1,24 +1,0 @@
-[
-    {
-        "version": "1.4.2-maykin-1.13.0",
-        "filterableColumns": [
-            "auteur",
-            "beschrijving",
-            "creatiedatum",
-            "informatieobjecttype",
-            "titel",
-            "vertrouwelijkheidaanduiding",
-            "trefwoorden"
-        ],
-        "sortableColumns": [
-            "auteur",
-            "bestandsomvang",
-            "creatiedatum",
-            "formaat",
-            "status",
-            "titel",
-            "vertrouwelijkheidaanduiding"
-        ],
-        "supportsTrefwoorden": true
-    }
-]
