+++ plugin/src/main/kotlin/com/ritense/plugin/domain/PluginConfigurationId.kt
@@ -31,10 +31,6 @@ data class PluginConfigurationId(
 
     companion object {
 
-        fun existingId(id: String): PluginConfigurationId {
-            return existingId(UUID.fromString(id))
-        }
-
         fun existingId(id: UUID): PluginConfigurationId {
             return PluginConfigurationId(id)
         }
