+++ zgw/documenten-api/src/main/kotlin/com/ritense/documentenapi/client/PatchDocumentRequest.kt
@@ -36,8 +36,7 @@ class PatchDocumentRequest(
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     val verzenddatum: LocalDate? = null,
     val indicatieGebruiksrecht: Boolean? = false,
-    val trefwoorden: List<String>? = null,
-    var lock: String? = null
+    var lock : String? = null
 ) {
     constructor(modifyDocumentRequest: ModifyDocumentRequest) : this(
         modifyDocumentRequest.creatiedatum,
@@ -49,7 +48,6 @@ class PatchDocumentRequest(
         modifyDocumentRequest.beschrijving,
         modifyDocumentRequest.ontvangstdatum,
         modifyDocumentRequest.verzenddatum,
-        modifyDocumentRequest.indicatieGebruiksrecht,
-        modifyDocumentRequest.trefwoorden,
+        modifyDocumentRequest.indicatieGebruiksrecht
     )
 }
