+++ zgw/documenten-api/src/main/kotlin/com/ritense/documentenapi/web/rest/ZgwDocumentTrefwoordResource.kt
@@ -27,7 +27,6 @@ import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
@@ -72,15 +71,6 @@ class ZgwDocumentTrefwoordResource(
         @PathVariable(name = "trefwoord") trefwoord: String,
     ): ResponseEntity<Page<ZgwDocumentTrefwoord>> {
         zgwDocumentTrefwoordService.deleteTrefwoord(caseDefinitionName, trefwoord)
-        return ResponseEntity.noContent().build()
-    }
-
-    @DeleteMapping("/management/v1/case-definition/{caseDefinitionName}/zgw-document/trefwoord")
-    fun deleteTrefwoorden(
-        @PathVariable(name = "caseDefinitionName") caseDefinitionName: String,
-        @RequestBody trefwoorden: List<String>,
-    ): ResponseEntity<Page<ZgwDocumentTrefwoord>> {
-        zgwDocumentTrefwoordService.deleteTrefwoorden(caseDefinitionName, trefwoorden)
         return ResponseEntity.noContent().build()
     }
 }
