--- zgw/documenten-api/src/main/kotlin/com/ritense/documentenapi/web/rest/dto/DocumentSearchRequest.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright 2015-2024 Ritense BV, the Netherlands.
- *
- * Licensed under EUPL, Version 1.2 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" basis,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ritense.documentenapi.web.rest.dto
-
-import com.fasterxml.jackson.annotation.JsonIgnore
-import java.net.URI
-import java.time.LocalDate
-
-data class DocumentSearchRequest(
-    val informatieobjecttype: String? = null,
-    val titel: String? = null,
-    val vertrouwelijkheidaanduiding: String? = null,
-    val creatiedatumFrom: LocalDate? = null,
-    val creatiedatumTo: LocalDate? = null,
-    val auteur: String? = null,
-    val trefwoorden: List<String>? = null,
-    @JsonIgnore val zaakUrl: URI? = null
-)
