--- zgw/documenten-api/src/main/resources/config/documenten-api/base-v1.zgw-documenten-api-version.json
@@ -1,46 +1,0 @@
-[
-    {
-        "version": "1.5.0",
-        "supportsTrefwoorden": true
-    },
-    {
-        "version": "1.4.3",
-        "supportsTrefwoorden": true
-    },
-    {
-        "version": "1.4.2",
-        "supportsTrefwoorden": true
-    },
-    {
-        "version": "1.4.1",
-        "supportsTrefwoorden": true
-    },
-    {
-        "version": "1.4.0",
-        "supportsTrefwoorden": true
-    },
-    {
-        "version": "1.3.0"
-    },
-    {
-        "version": "1.2.3"
-    },
-    {
-        "version": "1.2.2"
-    },
-    {
-        "version": "1.2.1"
-    },
-    {
-        "version": "1.2.0"
-    },
-    {
-        "version": "1.1.0"
-    },
-    {
-        "version": "1.0.1"
-    },
-    {
-        "version": "1.0.0"
-    }
-]
