+++ zgw/documenten-api/src/main/resources/config/liquibase/documenten-api-master.xml
@@ -22,8 +22,6 @@
 
     <property name="intType" value="int" dbms="mysql"/>
     <property name="intType" value="integer" dbms="postgresql"/>
-    <property name="jsonType" value="JSON" dbms="mysql"/>
-    <property name="jsonType" value="JSONB" dbms="postgresql"/>
 
     <include file="changelog/20240221-add-documente-api-list-column.xml" relativeToChangelogFile="true"/>
     <include file="changelog/20240222-add-documenten-api-trefwoorden.xml" relativeToChangelogFile="true"/>
