+++ zgw/documenten-api/src/test/kotlin/com/ritense/documentenapi/DocumentenApiPluginFactoryTest.kt
@@ -18,8 +18,6 @@ package com.ritense.documentenapi
 
 import com.fasterxml.jackson.databind.node.ObjectNode
 import com.ritense.documentenapi.client.DocumentenApiClient
-import com.ritense.documentenapi.service.DocumentDeleteHandler
-import com.ritense.documentenapi.service.DocumentenApiVersionService
 import com.ritense.plugin.domain.PluginConfiguration
 import com.ritense.plugin.domain.PluginConfigurationId
 import com.ritense.plugin.domain.PluginDefinition
@@ -42,11 +40,9 @@ internal class DocumentenApiPluginFactoryTest {
         val pluginService: PluginService = mock()
         val client: DocumentenApiClient = mock()
         val storageService: TemporaryResourceStorageService = mock()
+        val applicationEventPublisher = mock<ApplicationEventPublisher>()
         val authentication = mock<DocumentenApiAuthentication>()
-        val applicationEventPublisher: ApplicationEventPublisher = mock()
-        val objectMapper = MapperSingleton.get()
-        val documentDeleteHandlers: List<DocumentDeleteHandler> = mock()
-        val documentenApiVersionService: DocumentenApiVersionService = mock()
+
         whenever(pluginService.createInstance(any<PluginConfigurationId>())).thenReturn(authentication)
         whenever(pluginService.getObjectMapper()).thenReturn(MapperSingleton.get())
 
@@ -86,9 +82,8 @@ internal class DocumentenApiPluginFactoryTest {
             client,
             storageService,
             applicationEventPublisher,
-            objectMapper,
-            documentDeleteHandlers,
-            documentenApiVersionService,
+            MapperSingleton.get(),
+            listOf()
         )
 
         val plugin = factory.create(configuration)
