--- zgw/documenten-api/src/test/resources/config/pbac/all.permission.json
@@ -1,10 +1,0 @@
-{
-    "changesetId": "document-definitions-v1",
-    "permissions": [
-        {
-            "resourceType": "com.ritense.document.domain.impl.JsonSchemaDocumentDefinition",
-            "action": "view",
-            "roleKey": "ROLE_USER"
-        }
-    ]
-}
