--- zgw/documenten-api/src/test/resources/config/pbac/all.role.json
@@ -1,7 +1,0 @@
-{
-    "changesetId": "roles-v1",
-    "roles": [
-        "ROLE_USER",
-        "ROLE_ADMIN"
-    ]
-}
