+++ zgw/src/main/kotlin/com/ritense/zgw/ClientTools.kt
@@ -70,10 +70,5 @@ class ClientTools {
                 .path(uri.path)
                 .port(uri.port)
         }
-
-        fun UriBuilder.optionalQueryParam(name: String, value: Any?): UriBuilder {
-            value?.let { this.queryParam(name, it) }
-            return this
-        }
     }
 }
