--- zgw/zaken-api/src/main/kotlin/com/ritense/zakenapi/event/ZaakResultaatViewed.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright 2015-2024 Ritense BV, the Netherlands.
- *
- * Licensed under EUPL, Version 1.2 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" basis,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ritense.zakenapi.event
-
-import com.fasterxml.jackson.databind.node.ObjectNode
-import com.ritense.outbox.domain.BaseEvent
-
-class ZaakResultaatViewed(zaakResultaat: ObjectNode) : BaseEvent(
-    type = "com.ritense.gzac.zrc.resultaat.viewed",
-    resultType = "com.ritense.zakenapi.domain.ZaakResultaat",
-    resultId = null,
-    result = zaakResultaat
-)
