+++ zgw/zaken-api/src/main/kotlin/com/ritense/zakenapi/security/ZakenApiHttpSecurityConfigurer.kt
@@ -31,7 +31,6 @@ class ZakenApiHttpSecurityConfigurer : HttpSecurityConfigurer {
         try {
             http.authorizeHttpRequests { requests ->
                 requests.requestMatchers(antMatcher(GET, "/api/v1/zaken-api/document/{documentId}/files")).authenticated()
-                    .requestMatchers(antMatcher(GET, "/api/v2/zaken-api/document/{documentId}/files")).authenticated()
                     .requestMatchers(antMatcher(GET, "/api/v1/zaken-api/document/{documentId}/zaak")).authenticated()
                     .requestMatchers(antMatcher(GET, "/api/management/v1/zaak-type-link/{documentDefinitionName}")).hasAuthority(ADMIN)
                     .requestMatchers(antMatcher(GET, "/api/management/v1/zaak-type-link/process/{processDefinitionKey}")).hasAuthority(ADMIN)
