+++ zgw/zaken-api/src/main/kotlin/com/ritense/zakenapi/service/ZaakDocumentService.kt
@@ -18,19 +18,6 @@ package com.ritense.zakenapi.service
 
 import com.ritense.catalogiapi.service.CatalogiService
 import com.ritense.documentenapi.DocumentenApiPlugin
-import com.ritense.documentenapi.client.DocumentInformatieObject
-import com.ritense.documentenapi.domain.DocumentenApiColumnKey
-import com.ritense.documentenapi.domain.DocumentenApiColumnKey.AUTEUR
-import com.ritense.documentenapi.domain.DocumentenApiColumnKey.CREATIEDATUM
-import com.ritense.documentenapi.domain.DocumentenApiColumnKey.INFORMATIEOBJECTTYPE
-import com.ritense.documentenapi.domain.DocumentenApiColumnKey.TITEL
-import com.ritense.documentenapi.domain.DocumentenApiColumnKey.TREFWOORDEN
-import com.ritense.documentenapi.domain.DocumentenApiColumnKey.VERTROUWELIJKHEIDAANDUIDING
-import com.ritense.documentenapi.domain.DocumentenApiVersion
-import com.ritense.documentenapi.service.DocumentenApiService
-import com.ritense.documentenapi.service.DocumentenApiVersionService
-import com.ritense.documentenapi.web.rest.dto.DocumentSearchRequest
-import com.ritense.documentenapi.web.rest.dto.DocumentenApiDocumentDto
 import com.ritense.documentenapi.web.rest.dto.RelatedFileDto
 import com.ritense.plugin.domain.PluginConfiguration
 import com.ritense.plugin.service.PluginService
@@ -39,21 +26,15 @@ import com.ritense.zakenapi.ZakenApiPlugin
 import com.ritense.zakenapi.domain.ZaakInformatieObject
 import com.ritense.zakenapi.domain.ZaakResponse
 import com.ritense.zakenapi.link.ZaakInstanceLinkNotFoundException
-import org.springframework.data.domain.Page
-import org.springframework.data.domain.PageImpl
-import org.springframework.data.domain.Pageable
 import org.springframework.transaction.annotation.Transactional
 import java.net.URI
 import java.util.UUID
-import kotlin.math.min
 
 @Transactional
 class ZaakDocumentService(
     private val zaakUrlProvider: ZaakUrlProvider,
     private val pluginService: PluginService,
     private val catalogiService: CatalogiService,
-    private val documentenApiService: DocumentenApiService,
-    private val documentenApiVersionService: DocumentenApiVersionService,
 ) {
 
     fun getInformatieObjectenAsRelatedFiles(documentId: UUID): List<RelatedFileDto> {
@@ -70,65 +51,10 @@ class ZaakDocumentService(
             .map { getRelatedFiles(it) }
     }
 
-    fun getInformatieObjectenAsRelatedFilesPage(
-        documentId: UUID,
-        documentSearchRequest: DocumentSearchRequest,
-        pageable: Pageable,
-    ): Page<DocumentenApiDocumentDto> {
-        val zaakUri = zaakUrlProvider.getZaakUrl(documentId)
-        val version = documentenApiVersionService.getVersionByDocumentId(documentId)
-        check(documentSearchRequest.informatieobjecttype != null, INFORMATIEOBJECTTYPE, version)
-        check(documentSearchRequest.titel != null, TITEL, version)
-        check(documentSearchRequest.vertrouwelijkheidaanduiding != null, VERTROUWELIJKHEIDAANDUIDING, version)
-        check(documentSearchRequest.creatiedatumFrom != null, CREATIEDATUM, version)
-        check(documentSearchRequest.creatiedatumTo != null, CREATIEDATUM, version)
-        check(documentSearchRequest.auteur != null, AUTEUR, version)
-        check(documentSearchRequest.trefwoorden != null, TREFWOORDEN, version)
-        if (!version.supportsTrefwoorden) {
-            require(documentSearchRequest.trefwoorden == null) {
-                "Unsupported field 'trefwoorden' on Documenten API version '$version'"
-            }
-        }
-        return if (version.supportsFilterableColumns() && version.supportsSortableColumns()) {
-            documentenApiService.getCaseInformatieObjecten(
-                documentId,
-                documentSearchRequest.copy(zaakUrl = zaakUri),
-                pageable
-            ).map { mapDocumentenApiDocument(it, version) }
-        } else {
-            val zakenApiPlugin = getZakenApiPlugin(zaakUri)
-            val documenten = zakenApiPlugin.getZaakInformatieObjecten(zaakUri)
-                .map { mapDocumentenApiDocument(it, version) }
-            return toPage(documenten, pageable)
-        }
-    }
-
-    private fun check(shouldCheck: Boolean, columnKey: DocumentenApiColumnKey, version: DocumentenApiVersion) {
-        if (shouldCheck) {
-            val fieldName = columnKey.name.lowercase()
-            check(version.filterableColumns.contains(fieldName)) {
-                "Unsupported filter '$fieldName' on Documenten API with version $version"
-            }
-        }
-    }
-
-    private fun <T> toPage(list: List<T>, pageable: Pageable): Page<T> {
-        val startIndex = pageable.offset.toInt()
-        val endIndex = min(startIndex + pageable.pageSize, list.size)
-        return PageImpl(list.subList(startIndex, endIndex), pageable, list.size.toLong())
-    }
-
     private fun getRelatedFiles(zaakInformatieObject: ZaakInformatieObject): RelatedFileDto {
         val pluginConfiguration = getDocumentenApiPluginByInformatieobjectUrl(zaakInformatieObject.informatieobject)
         val plugin = pluginService.createInstance(pluginConfiguration) as DocumentenApiPlugin
         val informatieObject = plugin.getInformatieObject(zaakInformatieObject.informatieobject)
-        return mapRelatedFile(informatieObject, pluginConfiguration)
-    }
-
-    private fun mapRelatedFile(
-        informatieObject: DocumentInformatieObject,
-        pluginConfiguration: PluginConfiguration
-    ): RelatedFileDto {
         return RelatedFileDto(
             fileId = UUID.fromString(informatieObject.url.path.substringAfterLast("/")),
             fileName = informatieObject.bestandsnaam,
@@ -150,74 +76,6 @@ class ZaakDocumentService(
             confidentialityLevel = informatieObject.vertrouwelijkheidaanduiding?.key,
             version = informatieObject.versie,
             indicationUsageRights = informatieObject.indicatieGebruiksrecht
-        )
-    }
-
-    private fun mapDocumentenApiDocument(
-        informatieObject: DocumentInformatieObject,
-        version: DocumentenApiVersion,
-    ): DocumentenApiDocumentDto {
-        val pluginConfiguration = getDocumentenApiPluginByInformatieobjectUrl(informatieObject.url)
-        val trefwoorden = if (version.supportsTrefwoorden) {
-            informatieObject.trefwoorden
-        } else {
-            listOf()
-        }
-        return DocumentenApiDocumentDto(
-            fileId = UUID.fromString(informatieObject.url.path.substringAfterLast("/")),
-            pluginConfigurationId = pluginConfiguration.id.id,
-            bestandsnaam = informatieObject.bestandsnaam,
-            bestandsomvang = informatieObject.bestandsomvang,
-            creatiedatum = informatieObject.creatiedatum.atStartOfDay(),
-            auteur = informatieObject.auteur,
-            titel = informatieObject.titel,
-            status = informatieObject.status?.key,
-            taal = informatieObject.taal,
-            identificatie = informatieObject.identificatie,
-            beschrijving = informatieObject.beschrijving,
-            informatieobjecttype = getInformatieobjecttypeByUri(informatieObject.informatieobjecttype),
-            trefwoorden = trefwoorden,
-            formaat = informatieObject.formaat,
-            verzenddatum = informatieObject.verzenddatum,
-            ontvangstdatum = informatieObject.ontvangstdatum,
-            vertrouwelijkheidaanduiding = informatieObject.vertrouwelijkheidaanduiding?.key,
-            versie = informatieObject.versie,
-            indicatieGebruiksrecht = informatieObject.indicatieGebruiksrecht
-        )
-    }
-
-    private fun mapDocumentenApiDocument(
-        zaakInformatieObject: ZaakInformatieObject,
-        version: DocumentenApiVersion,
-    ): DocumentenApiDocumentDto {
-        val pluginConfiguration = getDocumentenApiPluginByInformatieobjectUrl(zaakInformatieObject.informatieobject)
-        val plugin = pluginService.createInstance(pluginConfiguration) as DocumentenApiPlugin
-        val informatieObject = plugin.getInformatieObject(zaakInformatieObject.informatieobject)
-        val trefwoorden = if (version.supportsTrefwoorden) {
-            informatieObject.trefwoorden
-        } else {
-            listOf()
-        }
-        return DocumentenApiDocumentDto(
-            fileId = UUID.fromString(informatieObject.url.path.substringAfterLast("/")),
-            pluginConfigurationId = pluginConfiguration.id.id,
-            bestandsnaam = informatieObject.bestandsnaam,
-            bestandsomvang = informatieObject.bestandsomvang,
-            creatiedatum = informatieObject.creatiedatum.atStartOfDay(),
-            auteur = informatieObject.auteur,
-            titel = informatieObject.titel,
-            status = informatieObject.status?.key,
-            taal = informatieObject.taal,
-            identificatie = informatieObject.identificatie,
-            beschrijving = informatieObject.beschrijving,
-            informatieobjecttype = getInformatieobjecttypeByUri(informatieObject.informatieobjecttype),
-            trefwoorden = trefwoorden,
-            formaat = informatieObject.formaat,
-            verzenddatum = informatieObject.verzenddatum,
-            ontvangstdatum = informatieObject.ontvangstdatum,
-            vertrouwelijkheidaanduiding = informatieObject.vertrouwelijkheidaanduiding?.key,
-            versie = informatieObject.versie,
-            indicatieGebruiksrecht = informatieObject.indicatieGebruiksrecht
         )
     }
 
@@ -257,15 +115,6 @@ class ZaakDocumentService(
         )
             ?: throw IllegalStateException("Missing plugin configuration of type '${ZakenApiPlugin.PLUGIN_KEY}' for url '$url'")
         return plugin.getZaak(url)
-    }
-
-    private fun getZakenApiPlugin(zaakUri: URI): ZakenApiPlugin {
-        return checkNotNull(
-            pluginService.createInstance(
-                ZakenApiPlugin::class.java,
-                ZakenApiPlugin.findConfigurationByUrl(zaakUri)
-            )
-        ) { "Could not find ${ZakenApiPlugin::class.simpleName} configuration for zaak with url: $zaakUri" }
     }
 
 }
