--- zgw/zaken-api/src/test/resources/config/form/form-with-zaakresultaat-fields.json
@@ -1,14 +1,0 @@
-{
-    "display": "form",
-    "components": [
-        {
-            "key": "zaakResultaat",
-            "type": "textfield",
-            "label": "Zaak resultaat",
-            "properties": {
-                "sourceKey": "zaakresultaat:omschrijving"
-            },
-            "input": true
-        }
-    ]
-}
